added new class EnterTextExamples.cs (extracted tests from ApiExamples.cs)
