+++ src/Coypu.AcceptanceTests/Examples/ApiExamples.cs
@@ -1,5 +1,4 @@
 ﻿using System;
-using System.IO;
 using Coypu.Drivers;
 using Coypu.Drivers.Selenium;
 using Coypu.NUnit.Matchers;
@@ -163,47 +162,6 @@ namespace Coypu.AcceptanceTests.Examples
         }
 
         [Test]
-        public void FillIn_file_example()
-        {
-            const string someLocalFile = @"local.file";
-            try
-            {
-                var directoryInfo = new DirectoryInfo(Path.GetTempPath());
-                var fullPath = Path.Combine(directoryInfo.FullName, someLocalFile);
-                using (File.Create(fullPath)) { }
-
-                Browser.FillIn("forLabeledFileFieldId")
-                       .With(fullPath);
-                var findAgain = Browser.FindField("forLabeledFileFieldId");
-                Assert.That(findAgain.Value, Does.EndWith(someLocalFile));
-            }
-            finally
-            {
-                File.Delete(someLocalFile);
-            }
-        }
-
-        [Test]
-        public void FillInWith_element_example()
-        {
-            Browser.FindField("scope2ContainerLabeledTextInputFieldId")
-                   .FillInWith("New text input value");
-            Assert.That(Browser.FindField("scope2ContainerLabeledTextInputFieldId")
-                               .Value,
-                        Is.EqualTo("New text input value"));
-        }
-
-        [Test]
-        public void FillInWith_example()
-        {
-            Browser.FillIn("scope2ContainerLabeledTextInputFieldId")
-                   .With("New text input value");
-            Assert.That(Browser.FindField("scope2ContainerLabeledTextInputFieldId")
-                               .Value,
-                        Is.EqualTo("New text input value"));
-        }
-
-        [Test]
         public void Hover_example()
         {
             Assert.That(Browser.FindId("hoverOnMeTest")
@@ -305,16 +263,6 @@ namespace Coypu.AcceptanceTests.Examples
         }
 
         [Test]
-        public void SendKeys_example()
-        {
-            Browser.FindField("containerLabeledTextInputFieldId")
-                   .SendKeys(" - send these keys");
-            Assert.That(Browser.FindField("containerLabeledTextInputFieldId")
-                               .Value,
-                        Is.EqualTo("text input field two val - send these keys"));
-        }
-
-        [Test]
         public void ShowsAllCssInOrder_example()
         {
             Assert.That(Browser,
@@ -339,7 +287,6 @@ namespace Coypu.AcceptanceTests.Examples
                                                                                     "Me! Pick me!")));
         }
 
-
         [Test]
         public void ShowsContentContaining_example()
         {
@@ -350,7 +297,6 @@ namespace Coypu.AcceptanceTests.Examples
                                                                                         "a",
                                                                                         "list")));
         }
-
 
         [Test]
         public void ShowsCssContaining_example()
